move all helper object definitions to separate file
