+++ Photon-Tinker/Mesh/MeshSetup.swift
@@ -29,36 +29,6 @@ class MeshSetup {
     static let particleMeshTXCharacterisiticUUID: CBUUID = CBUUID(string: "6FA90003-5C4E-48A8-94F4-8030546F36FC")
 }
 
-enum MeshSetupErrorSeverity {
-    case Error //can't continue at this point, but retrying might help
-    case Fatal //can't continue and flow has to be restarted
-}
-
-struct MeshSetupPeripheralCredentials {
-    var name: String
-    var mobileSecret: String
-}
-
-struct MeshSetupDataMatrix {
-    var serialNumber: String
-    var mobileSecret: String
-
-    init?(dataMatrixString: String) {
-        let regex = try! NSRegularExpression(pattern: "([a-zA-Z0-9]{15})[ ]{1}([a-zA-Z0-9]{15})")
-        let nsString = dataMatrixString as NSString
-        let results = regex.matches(in: dataMatrixString, range: NSRange(location: 0, length: nsString.length))
-
-        if (results.count > 0) {
-            let arr = dataMatrixString.split(separator: " ")
-            serialNumber = String(arr[0])//"12345678abcdefg"
-            mobileSecret = String(arr[1])//"ABCDEFGHIJKLMN"
-        } else {
-            return nil
-        }
-    }
-}
-
-
 // TODO: should be globally reference not just for mesh
 extension ParticleDeviceType : CustomStringConvertible {
     public var description: String {
